+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+8.1
+---
+
+ * Add `JsonSchemaDumper` to dump JSON Schema from configuration node definitions
+ * Add `BaseNode::isNullable()` to check if a node accepts null as input
+
 8.0
 ---